Fixed bottombar invite popup
